Reserva livro e pedir login

+++ BiblioMovel/app/src/main/java/br/ufg/inf/level6/bibliomovel/models/LivroInteresse.kt
@@ -4,6 +4,6 @@ package br.ufg.inf.level6.bibliomovel.models
  * Created by erick on 07/07/18.
  */
 data class LivroInteresse(
-        val cod: Int,
+        val cod: Long,
         var posicao_na_fila: Int
 )
